test_mkdata() takes a string, convert it to a calendar time and back again.
string is inputed to strptime() which produce a time struct. The time struct is inputed to mktime() which creates
a calendar time. gmtime() is called with the calendar time to produce a time struct with time ZULU/GMT. The ZULU
time struct is asserted by producing a new string with strftime(), which is used in the assertion.